src/genai/whatsapp_bot/requirements.txt
Flask
twilio
openai
python-dotenv
gunicorn
